requirements.txt
fastapi
langchain
langchain-google-genai
langchain_core
langchain-community
uvicorn[standard]
python-dotenv
tqdm
youtube-transcript-api 
pytube 
tiktoken
pypdf
unstructured
python-pptx
docx2txt
networkx
pandas
xlrd
openpyxl
gdown
pytest
PyPDF2
psutil